Refactor StockActivity onCreate method
- Move drawer and activity loading to sample
