+++ opensrp-stock/src/main/java/org/smartregister/stock/activity/StockActivity.java
@@ -2,18 +2,10 @@ package org.smartregister.stock.activity;
 
 import android.os.Build;
 import android.os.Bundle;
-import android.view.Gravity;
-import android.view.View;
 import android.widget.GridView;
-import android.widget.LinearLayout;
-import android.widget.TextView;
 
 import androidx.annotation.RequiresApi;
-import androidx.appcompat.app.ActionBarDrawerToggle;
 import androidx.appcompat.app.AppCompatActivity;
-import androidx.appcompat.widget.Toolbar;
-import androidx.core.view.GravityCompat;
-import androidx.drawerlayout.widget.DrawerLayout;
 
 import com.google.android.material.navigation.NavigationView;
 
@@ -34,48 +26,6 @@ public abstract class StockActivity extends AppCompatActivity {
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_stock);
-
-        Toolbar toolbar = (Toolbar) findViewById(R.id.location_switching_toolbar);
-        toolbar.setNavigationOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                //return to previous activity
-                finish();
-            }
-        });
-
-        TextView nameInitials = (TextView) findViewById(R.id.name_inits);
-        nameInitials.setText(getLoggedInUserInitials());
-
-        final DrawerLayout drawerLayout = (DrawerLayout) findViewById(R.id.drawer_layout);
-        final NavigationView drawer = getNavigationView();
-        if (drawer != null) {
-            DrawerLayout.LayoutParams lp = new DrawerLayout.LayoutParams(
-                    getNavigationViewWith(), LinearLayout.LayoutParams.MATCH_PARENT);
-            lp.gravity = Gravity.START;
-            drawerLayout.addView(drawer, lp);
-            final ActionBarDrawerToggle toggle = new ActionBarDrawerToggle(
-                    this, drawerLayout, toolbar, R.string.navigation_drawer_open, R.string.navigation_drawer_close);
-            drawerLayout.setDrawerListener(toggle);
-            toggle.syncState();
-
-            drawer.setNavigationItemSelectedListener(getNavigationViewListener());
-        }
-        nameInitials.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                if (drawer != null && !drawerLayout.isDrawerOpen(GravityCompat.START)) {
-                    drawerLayout.openDrawer(GravityCompat.START);
-                } else
-                    finish();
-            }
-        });
-        setSupportActionBar(toolbar);
-        toolbar.setTitle("Stock Control");
-        if (getSupportActionBar() != null) {
-            getSupportActionBar().setDisplayHomeAsUpEnabled(false);
-        }
 
         stockGrid = (GridView) findViewById(R.id.stockgrid);
     }
